+++ api/src/certification/shared/domain/constants/certification-candidates-errors.js
