+++ api/src/certification/shared/domain/validators/certification-candidate-validator.js
@@ -1,7 +1,7 @@
 import JoiDate from '@joi/date';
 import BaseJoi from 'joi';
 const Joi = BaseJoi.extend(JoiDate);
-import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../../lib/domain/constants/certification-candidates-errors.js';
+import { CERTIFICATION_CANDIDATES_ERRORS } from '../constants/certification-candidates-errors.js';
 import { subscriptionSchema } from './subscription-validator.js';
 
 const BILLING_MODES = {
