+++ api/tests/certification/shared/unit/domain/services/certification-cpf-service_test.js
@@ -1,4 +1,4 @@
-import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../../../lib/domain/constants/certification-candidates-errors.js';
+import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../../../src/certification/shared/domain/constants/certification-candidates-errors.js';
 import {
   CpfBirthInformationValidation,
   getBirthInformation,
