+++ api/tests/unit/domain/models/CertificationCandidate_test.js
@@ -1,5 +1,5 @@
-import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../lib/domain/constants/certification-candidates-errors.js';
 import { Subscription } from '../../../../src/certification/enrolment/domain/models/Subscription.js';
+import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../src/certification/shared/domain/constants/certification-candidates-errors.js';
 import {
   CertificationCandidatePersonalInfoFieldMissingError,
   CertificationCandidatePersonalInfoWrongFormat,
